refactor(build): move copy sources script in the separate file
